chore: clippy cleanup sweep — 121+38 warnings/errors → 0 (slice 5.5)
Mirror of slice 5.4 for clippy. Drives workspace clippy from 121
warnings + 38 hidden deny-by-default errors to 0 errors + 1 cargo
config warning (the `[profile.release]` in pyde-crypto-wasm that
cargo ignores in workspace context — not emitted by clippy, so
`-D warnings` in CI won't catch it).

Unlocks task 048 fully: slice 5.6 can now wire
`cargo clippy --workspace -- -D warnings` into CI as a trivial
YAML addition.

Real bugs caught by the sweep:

- `node/src/node.rs::handle_decrypted_block_execution`: a `MutexGuard`
  from `state_w.smt_mut()` was being held across a `.await` inside
  the match body (on `receipts.write()`). Tokio's scheduler can move
  futures between threads; holding a std `Mutex` guard across an
  await is a well-known deadlock / UB pattern. Fixed by binding
  `execute_transaction`'s result BEFORE the match so the temporary
  guard drops at the semicolon.

- `node/tests/production_sim.rs::deploy_bench`: `if acc_idx < 10 {
  accounts[acc_idx].pk.as_bytes() } else { accounts[acc_idx].pk.as_bytes() }` —
  both arms identical (leftover from a refactor where deploy vs
  non-deploy accounts lived in separate collections). Collapsed to
  direct index.

- `node/src/node.rs::PostEventAction`: two `if` arms both returned
  `ContinueSync` — collapsed into `if A || B { ContinueSync } else
  { None }`.

- `tx/src/pipeline.rs`: `let success = success;` — dead rebinding
  from a refactor. Removed.

- `tx/src/multisig.rs`: `vec![((MAX_SIGNERS + 1))]` — extra parens
  from an earlier rewrite. Cleaned.

Workspace-level lint configuration:

Added `[workspace.lints.clippy]` in the root Cargo.toml with per-
crate opt-in via `[lints] workspace = true` in each member crate.
This documents the project's style decisions where clippy's
defaults don't match the codebase's established conventions:

  - `too_many_arguments`: consensus + pipeline functions naturally
    carry 8-12 primitive args (slot/base_fee/chain_id/validator/...).
    Packing into a struct adds indirection without improving
    readability.
  - `needless_range_loop`: crypto polynomial / share math reads
    more cleanly as `for i in 0..N { arr[i] = ... }` when the
    domain is fixed-size and multiple arrays share the index.
  - `type_complexity`: compiler + VM internals pass around nested
    generics where the shape IS the documentation.
  - `doc_lazy_continuation` / `empty_line_after_doc_comments`: doc
    style preferences that align with the existing codebase rather
    than clippy's current defaults.

Scoped (per-module / per-function) allows with explanatory comments:

  - `clippy::arc_with_non_send_sync` on test modules that capture
    raw `*const PydeSMT` pointers in closures (otic/codegen.rs test
    module; tx/pipeline.rs `e2e_rank_contract` test). Alternative
    would be switching `Vm::storage_backend` from Arc to Rc
    project-wide — a refactor that only matters if we ever
    parallelize VM execution.
  - `clippy::not_unsafe_ptr_arg_deref` at aot/src/host.rs module
    level. Every host function is `extern "C"` taking `*mut VmCtx`
    from the JIT. Marking them `unsafe extern "C" fn` would force
    `unsafe { }` at every JIT emission site — purely ergonomic
    friction; the generated assembly derefs unconditionally anyway.
    Module-level comment documents the JIT-side safety contract.
  - `clippy::eq_op` at otic/src/codegen.rs module level — patterns
    like `(14 & 0xF) << 2` deliberately keep the 4-bit mask to
    document the register-encoding field width.
  - `clippy::erasing_op` on one consensus/block.rs test asserting
    `0 / EPOCH_LENGTH == 0` (genesis sanity).
  - `clippy::cloned_ref_to_slice_refs` on block.rs + hotstuff.rs
    test modules (1-element `&[tx.clone()]` determinism asserts).
  - `clippy::only_used_in_recursion` on two otic/safety.rs methods
    that take `&self` for call-site symmetry with sibling methods
    but only pass it through to recursive calls.
  - `clippy::explicit_counter_loop` on one prop_integration test
    where `outer_nonce` tracks submitter tx-nonce (semantically
    distinct from the attempt index even though they happen to
    match).
  - `clippy::nonminimal_bool` on otic/typecheck.rs type-compat
    check — the chain-of-ANDs with per-line comments is what
    makes the compatibility matrix readable; clippy's De-Morgan
    rewrite loses that.

Mechanical fixes via cargo clippy --fix:

  - Redundant closures, same-type casts, `map_or` simplifications,
    `is_err()` over `while let Err(_) = ...`, `contains_key(&k)`
    over `get(&k).is_some()`, et cetera.
  - Applied across ~70 files, workspace-wide.

Files: 70 modified (including 15 Cargo.toml files with the new
`[lints]` block), +395 / -218 lines.
Workspace tests: 2322 passed, 0 failed (unchanged from pre-5.5).

+++ crates/node/src/chain.rs
@@ -41,7 +41,7 @@ impl ChainState {
     /// Advance chain head after processing a block.
     pub fn advance(&mut self, header: BlockHeader) {
         let slot = header.slot;
-        let epoch = slot / EPOCH_LENGTH as u64;
+        let epoch = slot / EPOCH_LENGTH;
 
         self.head_slot = slot;
         self.epoch = epoch;
@@ -86,7 +86,7 @@ mod tests {
     fn dummy_header(slot: u64, parent_hash: [u8; 32]) -> BlockHeader {
         BlockHeader {
             slot,
-            epoch: slot / EPOCH_LENGTH as u64,
+            epoch: slot / EPOCH_LENGTH,
             parent_hash,
             proposer: ZERO_ADDRESS,
             vrf_proof: vec![],
